Create project
This geospatial analysis project uses sf and tmap along with dplyr to highlight 'dangerous' and 'safe' neighborhoods in Detroit based on 2021 crime data. Then, utilizing a database on grocery store and bus stop locations, the analysis highlights which grocery stores are in what neighborhood, and how close grocery store are to bus stops.